1.Delete yarn.lock and install dependencies again 2.Move /profile endpoint from auth to user
