+++ src/controllers/auth.js
@@ -33,21 +33,20 @@ const loginUsers = async (req, res) => {
 };
 
 // eslint-disable-next-line consistent-return
-const getProfile = async (req, res) => {
-  try {
-    if (req.user) {
-      const currentUser = await UserModel.findById(req.user.userId);
-      return res
-        .setHeader('Content-Type', 'application/json')
-        .status(200)
-        .json(currentUser);
-    }
-  } catch (error) {
-    return res.status(403).json({ message: error.message });
-  }
-};
+// const getProfile = async (req, res) => {
+//   try {
+//     if (req.user) {
+//       const currentUser = await UserModel.findById(req.user.userId);
+//       return res
+//         .setHeader('Content-Type', 'application/json')
+//         .status(200)
+//         .json(currentUser);
+//     }
+//   } catch (error) {
+//     return res.status(403).json({ message: error.message });
+//   }
+// };
 
 module.exports = {
   loginUsers,
-  getProfile,
 };
