+++ src/routes/user.js
@@ -1,6 +1,8 @@
 const express = require('express');
-// const userController = require('../controllers/user');
+const userController = require('../controllers/user');
 
 const router = express.Router();
+
+router.get('/profile', userController.getProfile);
 
 module.exports = router;
